requirements.txt
argcomplete==3.6.2
bitarray==3.4.3
bitstring==4.3.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
ecdsa==0.19.1
esptool==4.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pyserial==3.5
PyYAML==6.0.2
reedsolo==1.7.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
